% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_layer.R
\name{$.soil_layer}
\alias{$.soil_layer}
\title{Cash soil_layer}
\usage{
\method{$}{soil_layer}(x, name)
}
\arguments{
\item{x}{A soil_layer}

\item{name}{Field name (character; unit length)}
}
\value{
Values of \code{name} in \code{x}
}
\description{
Cash soil_layer
}
