% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_profile.R
\name{$.soil_profile}
\alias{$.soil_profile}
\title{Cash field values of a soil_profile by name}
\usage{
\method{$}{soil_profile}(x, name)
}
\arguments{
\item{x}{A soil_profile}

\item{name}{Field name (character; unit length)}
}
\value{
Values of field name or error if field name does not exist.
}
\description{
Cash field values of a soil_profile by name
}
