% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_layer.R
\name{$<-.soil_layer}
\alias{$<-.soil_layer}
\title{Cash<- soil_layer}
\usage{
\method{$}{soil_layer}(x, name, i = NULL) <- value
}
\arguments{
\item{x}{A soil_layer}

\item{name}{Field name (character; unit length)}

\item{i}{Numeric or logical index}

\item{value}{New values}
}
\description{
Cash<- soil_layer
}
