% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_profile.R
\name{$<-.soil_profile}
\alias{$<-.soil_profile}
\title{Cash setter for field values in a soil_profile}
\usage{
\method{$}{soil_profile}(x, name, i = NULL) <- value
}
\arguments{
\item{x}{A soil_profile}

\item{name}{Field name (character; unit length)}

\item{i}{Numeric or logical index}

\item{value}{New values}
}
\value{
A soil_profile
}
\description{
Cash setter for field values in a soil_profile
}
