% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_layer.R
\name{default_metadata_soil_layer}
\alias{default_metadata_soil_layer}
\title{Get default soil_layer metadata}
\usage{
default_metadata_soil_layer()
}
\value{
A named list containing default field names for soil_layer objects
}
\description{
Get default soil_layer metadata
}
\examples{

# construct an empty soil_layer with default metadata

soil_layer(metadata = default_metadata_soil_layer())

}
