% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_profile.R
\name{default_metadata_soil_profile}
\alias{default_metadata_soil_profile}
\title{Get default soil_profile metadata}
\usage{
default_metadata_soil_profile()
}
\value{
A named list containing default field names for soil_profile objects
}
\description{
Get default soil_profile metadata
}
\examples{

# construct an empty soil_profile with default metadata

soil_profile(metadata = default_metadata_soil_profile())

}
