% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_layer.R
\name{.new_soil_layer}
\alias{.new_soil_layer}
\title{Internal constructor for \code{soil_layer} object}
\usage{
.new_soil_layer(
  x = list(pid = integer(), hid = integer(), geom = geovctrs::geo_rect()),
  metadata = list(pid = "pid", hid = "hid", ztop = "ztop", zbot = "zbot")
)
}
\arguments{
\item{x}{A named list containing, at a minimum, \code{pid}, \code{profile}, and \code{geom} named vectors.}
}
\value{
A \code{soil_layer}
}
\description{
Requires minimum 3 vectors (all of same length) in a named list \code{x}, with following \code{name : class}:
}
\details{
- \code{pid : character}
 - \code{hid : integer}
 - \code{geom : geovctrs::geo_rect}
}
\keyword{internal}
