% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_profile.R
\name{.new_soil_profile}
\alias{.new_soil_profile}
\title{Internal constructor for \code{soil_profile} object}
\usage{
.new_soil_profile(
  x = list(pid = integer(), profile = list(), geom = geovctrs::geo_xyz()),
  metadata = list(pid = "pid", hid = "hid", x = "x", y = "y", z = "z", ztop = "ztop",
    zbot = "zbot")
)
}
\arguments{
\item{x}{A named list containing, at a minimum, \code{pid}, \code{profile}, and \code{geom} named vectors.}
}
\value{
A \code{soil_profile}
}
\description{
Requires minimum 3 vectors (all of same length) in a named list \code{x}, with following \code{name : class}:
}
\details{
- \code{pid : character}
 - \code{profile : soilvctrs::soil_layer}
 - \code{geom : geovctrs::geo_xyz}
}
\keyword{internal}
