% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_layer.R
\name{format.soil_layer}
\alias{format.soil_layer}
\title{format.soil_layer}
\usage{
\method{format}{soil_layer}(x, ...)
}
\arguments{
\item{x}{A soil_layer}

\item{...}{Additional arguments [not used]}
}
\value{
Formatted soil_layer object
}
\description{
format.soil_layer
}
