% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_profile.R
\name{format.soil_profile}
\alias{format.soil_profile}
\title{format.soil_profile}
\usage{
\method{format}{soil_profile}(x, ...)
}
\arguments{
\item{x}{A soil_profile}

\item{...}{Additional arguments [not used]}
}
\value{
Formatted soil_profile object
}
\description{
format.soil_profile
}
