% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_layer.R
\name{is.soil_layer}
\alias{is.soil_layer}
\title{Does an R object inherit from soil_layer?}
\usage{
\method{is}{soil_layer}(x)
}
\arguments{
\item{x}{An object}
}
\value{
logical
}
\description{
Does an R object inherit from soil_layer?
}
