% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_profile.R
\name{is.soil_profile}
\alias{is.soil_profile}
\title{Does an R object inherit from soil_profile?}
\usage{
\method{is}{soil_profile}(x)
}
\arguments{
\item{x}{An object}
}
\value{
logical
}
\description{
Does an R object inherit from soil_profile?
}
