% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_layer.R
\name{plot.soil_layer.base}
\alias{plot.soil_layer.base}
\title{Default plot function for soil_layer -- sets axes etc.}
\usage{
\method{plot}{soil_layer.base}(x, xmax = length(x), ymax = 200, ...)
}
\arguments{
\item{x}{A soil_layer (generally from single soil_profile)}

\item{xmax}{Length (from 0) of X-axis (profile index axis)}

\item{ymax}{Length (from 0) in (negative Z direction) of plot Y-axis}

\item{...}{Additional arguments to \code{plot}}
}
\value{
A base graphics plot (to the current graphics device)
}
\description{
Default plot function for soil_layer -- sets axes etc.
}
