% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_layer.R
\name{plot.soil_layer}
\alias{plot.soil_layer}
\title{Rudimentary plotting method for soil_profile}
\usage{
\method{plot}{soil_layer}(
  x,
  base.plot.method = plot.soil_layer.base,
  xmax = length(x),
  ymax = 200,
  ...
)
}
\arguments{
\item{x}{A soil_layer}

\item{base.plot.method}{A function that can take x as an argument to set up plot pane. Default is \code{plot.soil_layer.base}.}

\item{xmax}{Length (from 0) of X-axis (profile index axis)}

\item{ymax}{Length (from 0) in (negative Z direction) of plot Y-axis}

\item{...}{Additional arguments to \code{plot(..., add=TRUE)} call for each layer}
}
\value{
A base graphics plot (to the current graphics device)
}
\description{
For inspection of geometry of a soil_layer
}
