% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_profile.R
\name{plot.soil_profile}
\alias{plot.soil_profile}
\title{Rudimentary plotting method for soil_profile}
\usage{
\method{plot}{soil_profile}(
  x,
  xmax = length(x),
  ymax = 200,
  base.plot.method = plot.soil_profile.base,
  ...
)
}
\arguments{
\item{x}{A soil_profile}

\item{xmax}{Length (from 0) of X-axis (profile index axis)}

\item{ymax}{Length (from 0) in (negative Z direction) of plot Y-axis}

\item{base.plot.method}{An R function that takes \code{x} as input and creates a base plot to draw profile geometry on.}

\item{...}{Additional arguments to \code{plot}}
}
\value{
A base graphics plot (to the current graphics device)
}
\description{
For inspection of geometry of a soil_profile
}
