% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_layer.R
\name{print.soil_layer}
\alias{print.soil_layer}
\title{print.soil_layer}
\usage{
\method{print}{soil_layer}(x, ...)
}
\arguments{
\item{x}{A soil_layer}

\item{...}{Additional arguments [not used]}
}
\value{
Print a soil_layer object
}
\description{
print.soil_layer
}
