% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_profile.R
\name{print.soil_profile}
\alias{print.soil_profile}
\title{print.soil_profile}
\usage{
\method{print}{soil_profile}(x, ...)
}
\arguments{
\item{x}{A soil_profile}

\item{...}{Additional arguments [not used]}
}
\value{
Print a soil_profile object
}
\description{
print.soil_profile
}
