% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_layer.R
\name{soil_layer}
\alias{soil_layer}
\title{Constructor for soil_layer object}
\usage{
soil_layer(
  layer = data.frame(pid = integer(0), hid = integer(0), ztop = integer(0), zbot =
    integer(0)),
  metadata = list(pid = "pid", hid = "hid", ztop = "ztop", zbot = "zbot")
)
}
\arguments{
\item{layer}{A data.frame-like object containing layer (horizon) data -- records with top and bottom depths}

\item{metadata}{Named list of field name aliases for profile ID, layer/horizon ID and layer top and bottom depth.}
}
\value{
A soil_layer
}
\description{
Constructor for soil_layer object
}
\details{
Metadata list specifications: \code{pid}: Field name containing profile ID in \code{site}; default: \code{"pid"}; \code{hid}: Field name containing layer ID in \code{horizons}; default: \code{"hid"}; \code{ztop}: Field name containing layer top depth in \code{horizons}; default: \code{"ztop"}; \code{zbot}: Field name containing layer bottom depth in \code{horizons}; default: \code{"zbot"}
}
