% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_profile.R
\name{soil_profile}
\alias{soil_profile}
\title{Constructor for soil_profile (from a list of soil_layers)}
\usage{
soil_profile(
  layer = data.frame(),
  site = data.frame(),
  metadata = list(pid = "pid", x = "x", y = "y", z = "z", hid = "hid", ztop = "ztop",
    zbot = "zbot")
)
}
\arguments{
\item{layer}{data.frame-like object containing site data (1 row = 1 layer; each becomes an element of a soil_layer)}

\item{site}{data.frame-like object containing site data (1 rows = 1 profiles; n-length soil_layer per profile)}

\item{metadata}{Named list of field name aliases for profile ID, X, Y and Z.}
}
\value{
A soil_profile
}
\description{
Constructor for soil_profile (from a list of soil_layers)
}
\details{
Metadata list specifications: \code{pid}: Field name containing profile ID in \code{site}; default: \code{"pid"}; \code{xpos}: Field name containing X-position in \code{site}; default: \code{"x"}; \code{ypos}: Field name containing Y-position in \code{site}; default: \code{"y"}; \code{zpos}: Field name containing Z-position in \code{site}; default: \code{"z"}
}
