% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_layer.R
\name{[.soil_layer}
\alias{[.soil_layer}
\title{Subset a soil_layer}
\usage{
\method{[}{soil_layer}(x, i = NULL, j = NULL)
}
\arguments{
\item{x}{A soil_layer}

\item{i}{Optional: Index positions (\code{1:length(x)} if not specified)}

\item{j}{Optional: field names for horizon-level attributess}
}
\value{
A subset of a soil_layer, by index \code{i}, and/or column name \code{j} to return a data.frame-like object
}
\description{
Using just the \code{i} index performs a simple subset of a soil_layer.
}
\details{
A single value in the \code{j} index will access the column values of that name (\code{j}).

If multiple Column names are specified with the \code{j} index, the result is a data.frame-like object.
}
