% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_profile.R
\name{[.soil_profile}
\alias{[.soil_profile}
\title{Subset a soil_profile}
\usage{
\method{[}{soil_profile}(x, i = NULL, j = NULL)
}
\arguments{
\item{x}{A soil_profile}

\item{i}{Optional: Index positions (\code{1:length(x)} if not specified)}

\item{j}{Optional: field names for site-level attributes}
}
\value{
A subset of a soil_profile, by index \code{i}, and/or column name \code{j} to return a data.frame-like object
}
\description{
Subset a soil_profile
}
