% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_layer.R
\name{[[.soil_layer}
\alias{[[.soil_layer}
\title{Subsub a soil_layer using a field name}
\usage{
\method{[[}{soil_layer}(x, name)
}
\arguments{
\item{x}{A soil_layer}

\item{name}{Field name (character; unit length)}
}
\value{
Values of field name or error if field name does not exist.
}
\description{
Subsub a soil_layer using a field name
}
